- Throw NotFoundException when try deleting unexisting resource - Delete Unnecessary getAll method, transitioning to search for sorting GET /products

+++ src/main/java/com/zamazor/market/modules/product/service/CategoryService.java
@@ -45,6 +45,9 @@ public class CategoryService {
 	}
 
 	public void delete(UUID id) {
+		if (!categoryRepository.existsById(id)) {
+			throw new CategoryNotFoundException("Category with id: " + id + " not found");
+		}
 		categoryRepository.deleteById(id);
 	}
 }
